Adds Put Post Delete methods to PizzaController
Added the put post and delete method to the PizzaController. Put allows the user to update a pizza via id, post allows the creation of a pizza, only a name and isGlutenFree boolean is required to create a pizza, the id is handled automatically. The delete command allows a user to delete an entry by its id