Implements #6
Created a fake login system by running a get command with a guid attached. Running the get command with a guid, correct guid returns nocontent and essentially "logs in" the user as a manager or CEO and allows the put and delete commands to be used. A failed attempt gives back a bad request but allows the user to continue as a regular employee