Implemented ability to upload to google cloud
Implemented a function that is used by calling get guid with a specific guid. When this is called it will grab the SavedList.json file in the working directory of the program and upload it to google cloud. Currently it uploads to the root directory of the bucket within google cloud. Hoping I can make it upload to a specific directory (i.e. quickstart-folder/ContosoPizza)