Fixed updatebynumber and updatebyname
Update by number and update by name were both working improperly after some testing. Neither name or number would work if you were changing the customers name, or number. Now both functions will grab the input name and input number and pass it to the API to check for the customer before modifying the data. This allows for a customer to be updated to where every field is different, as if you were adding a brand new customer