Added additional support for MVC requests
Implemented an update by name from MVC method that can be used by the python MVC program. Also implemented a catch when requesting a new customer to check if they are being added with a duplicate phone number, and send an accepted response, which means it was accepted but nothing was done with the request.