Modified phone number types
Changed phone numbers from int to long as it was causing issues. Found out through testing in my python MVC that once you enter a phone number with a greater area code than 214, it would cause an error of using an invalid type, due to the size limitation of an int.